Add files via upload
SEEDO is an online tailoring/stitching website which provides a door-to-door service to the customer. SEEDO is a complete and custom, stitching service application designed to offer quality tailored garments, clothing alterations for men, women and children; with promised hassle-free tailoring experience at affordable prices with free pick-up and delivery services.
This  custom tailoring websites enable users to select fabric, design it as per their style, and get the stitched piece of clothing delivered to their doorstep.
This website provides services in Kolkata for now. Further, it would expand its services to other cities in India.
This website works on the concept of pick up, stitch it and deliver. When the customer books an appointment, an experienced tailor is sent to the doorstep of the customer who helps the customer in designing their clothes to be tailored(ready to stitch), their requirements of styles, fabrics and different elements of their own choice is noted by the tailor and measurements are taken. The customer can also pick their style from the website catalogue or instruct the tailor according to their preferences. A customer’s best fitted garment is taken as a sample to get the best fit and comfort of the customer. Then the stitching process takes place and it is delivered to the customer at their doorstep without letting the customer to come to the retail outlet. This eases out the customer’s travel charges, time and provides home service without any hassle.
The customer can contact the website via phone number or email given on the website for any doubts, keeping in update about the shipment details etc. SEEDO delivers the product within 7 days and if not able to give the product on time, it compensates for the same by returning the money. SEEDO has an express delivery service for faster delivery for which extra charges are involved.
After receiving the garment, the customer can contact SEEDO for any sort of further alterations needed. It also registers complaints from the customer , if the customer is not satisfied with the product or found any sort of defects in the product(maybe detailing is not as per their preference given, or any major defects found) then the customer can return the product. The delivery boy would pick up the product for return, check it and the money will be returned to the customer.
SEEDO provides payment modes such as cash on delivery, card(VISA, MASTERCARD), net banking etc. The customer needs to pay after getting the product at the doorstep.